refactor withdrawals; add rollover page
